refactor: move proto serialization into its own module
A leaf that neither voice nor telemetry sits above, so both can share one
encoder. Verbatim otherwise.
